requirements.txt
# Base requirements
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9 
djangorestframework-gis>=1.0
django-filter>=24.0 
gunicorn>=21.0
drf-spectacular>=0.27
geopy>=2.4